requirements.txt
pandas==2.1.4
requests-oauthlib==1.3.1
python-dotenv==1.0.0
xlsxwriter==3.1.9
schwabdev==1.0.0